pyproject.toml
[project]
name = "crud-analytics-api"
version = "0.1.0"
description = ""
authors = [
    {name = "isqfc",email = "iffcorreia2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.12.11"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
faker = "^37.6.0"

[tool.ruff]
extend-exclude = ['migrations']
line-length = 79

[tool.ruff.lint]
select = ['E', 'F', 'I', 'W', 'PT', 'PL']
preview = true

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'

[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff check --fix'
run = 'fastapi dev crud_analytics_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=crud_analytics_api'